kata/tdd/drf-developer/gptsolution/requirements.txt
django
django-filter
djangorestframework
drf-yasg
markdown
